Initial stab at the problem
Currently we don't serialise List<OrderSummary>.

We have no tests requiring the username, but I thought I'd chuck it in
anyway. The problem description only requires the username for use in
the textual summary it might be overkill to wrap it up in an object.

I'm not very happy with how we perform ordering of the different
orderings. Currently the OrderSummariser is responsible for it, but that
couples it quite tightly to the Order object hierarchy--perhaps we
should shift the responsibility onto the Orders themselves so that they
can order themselves. I'm hesitant to implement Comparable as you might
want to compare them in different ways for different purposes.